requirements.txt

openai>=1.0.0
requests
google-generativeai
playwright
playwright-stealth
